+++ modules/tcp/TCP.h
@@ -15,8 +15,8 @@
  */
 #pragma once
 
+#include <stdint.h>
 #include <sys/socket.h>
-#include <sys/types.h> /* See NOTES */
 
 #include <string>
 
@@ -45,10 +45,10 @@ class TCP {
     TCP(const std::string &host, const ConnectInfo &ConnectInfo);
     virtual ~TCP(void);
 
-    void Send(const void *data, size_t &szData);
-    void SendSizedData(const void *data, size_t &szData);
-    int Recv(void *data, size_t &szData);
-    int RecvSizedData(void **data, size_t &szData);
+    void Send(const void *data, int32_t &szData);
+    void SendSizedData(const void *data, int32_t &szData);
+    int Recv(void *data, int32_t &szData);
+    int RecvSizedData(void **data, int32_t &szData);
     int GetHandle(void);
     unsigned short GetPort(void);
     void GetPeerInfo(std::string &host, unsigned short &port);
@@ -56,11 +56,11 @@ class TCP {
   private:
     TCP(int handle, sockaddr *addr, socklen_t addrlen, const ConnectInfo &connect_info);
     void SetupOptions(const ConnectInfo &connect_info);
-    int HandleZeroMsg(void **data, size_t &data_size);
-    void SendSizedDataNormal(const void *data, size_t &data_size);
-    int RecvSizedDataNormal(void **data, size_t &data_size);
-    void SendSizedDataSecure(const void *data, size_t &data_size);
-    int RecvSizedDataSecure(void **data, size_t &data_size);
+    int HandleZeroMsg(void **data, int32_t &data_size);
+    void SendSizedDataNormal(const void *data, int32_t &data_size);
+    int RecvSizedDataNormal(void **data, int32_t &data_size);
+    void SendSizedDataSecure(const void *data, int32_t &data_size);
+    int RecvSizedDataSecure(void **data, int32_t &data_size);
 
     int handle;
     socklen_t addrlen;
